Implements type promotion in probability computations.
This addresses numerical issues by conducting probability computations with a minimum dtype. In addition, some special functions are not implemented for less than float32, like the normal CDF (error function). So this is necessary anyway.

Also adds a few more type annotations to the API, and makes the API of the soft rounding ops more consistent.

PiperOrigin-RevId: 533423254
Change-Id: Ie3d7e30f3b9670d04e9c9a171e4cd95d1d2f6e3c