Adds scale_param.
This helper function parameterizes a scale parameter of a distribution. Main use would be in a conditional entropy model, for taking a neural network output and outputting a positive scale parameter. It limits the scale to a finite range, which prevents numerical issues, and sets up the model for quantizing the scale parameter, which would be needed for range coding.

Usage should look something like this:

```
class ConditionalEntropyModel(cdx.ems.DistributionEntropyModel):
  param: Array

  @property
  def distribution(self):
    loc = self.param[..., 0::2]
    scale = cdx.ems.scale_param(self.param[..., 1::2], 20)
    return tfp.distributions.Normal(loc=loc, scale=scale)
```

This commit also adds upper_limit/lower_limit, akin to upper_bound/lower_bound in TFC.

PiperOrigin-RevId: 538887982
Change-Id: I17a4b19e9fd9a2644d2705a56da2e686577a3e87